Refactor cdq into pkg (#76)
* Refactor SubFunction code into public package for usability outside this project

* Inlcude wrapper comments for subfunctions API calls
