+++ internal/deviceplugin/poolManager.go
@@ -82,10 +82,10 @@ func (pm *PoolManager) Init(config PoolConfig) error {
 	pm.BpfHandler = bpf.NewHandler()
 	pm.NetHandler = networking.NewHandler()
 
-        if err := pm.startGRPC(); err != nil {
-                return err
-        }
-        logging.Infof("Pool "+pm.DevicePrefix+"/%s started serving", pm.Name)
+	if err := pm.startGRPC(); err != nil {
+		return err
+	}
+	logging.Infof("Pool "+pm.DevicePrefix+"/%s started serving", pm.Name)
 
 	if err := pm.registerWithKubelet(); err != nil {
 		return err
